stop using removed impl

+++ src/iterators/integers.rs
@@ -311,7 +311,7 @@ impl Iterator for RandomRangeInteger {
     type Item = Integer;
 
     fn next(&mut self) -> Option<Integer> {
-        Some(random_natural_below(&mut self.rng, &self.diameter_plus_one) + &self.a)
+        Some(Integer::from(random_natural_below(&mut self.rng, &self.diameter_plus_one)) + &self.a)
     }
 }
